fix: JWT 토큰 수명 설정 단위 오류 수정 (#29)
yml의 숫자는 밀리초 의도로 작성됐는데 코드가 초로 읽어 ×1000 하고 있었다.
액세스 토큰이 416일, 리프레시 토큰이 27년간 유효했다.

소비 지점 네 곳이 일관되게 초로 다뤄 코드 내부에는 모순이 없었고,
그래서 지금까지 아무 증상 없이 동작했다. signOut()은 Redis의 리프레시
토큰만 지우므로 자체 완결형인 액세스 토큰은 로그아웃 후에도 exp까지 통과했다.

TokenProperties.expiration을 Duration으로 바꿔 단위를 타입에 새기고,
yml은 10h / 10d로 표기한다. 소비 지점마다 toMillis() / toSeconds()를
명시해 초·밀리초 선택이 코드에 드러나게 했다.

RefreshToken.timeToLive는 Redis @TimeToLive가 초를 받으므로 Long으로 넓혔다.

테스트 3개는 각각 다른 지점을 지킨다.
- JwtProviderKoTest: 토큰의 iat↔exp 간격. 벽시계로 재면 발급까지의
  경과분이 섞여 초 단위 버림에서 흔들린다
- AuthenticationServiceKoTest: Redis TTL과 쿠키 Max-Age. 단위를 잘못
  고르기 가장 쉬운 자리다
- JwtPropertiesKoTest: 운영 application-common.yml을 직접 읽는다.
  나머지 테스트는 test/resources/application.yml만 보므로 복제본인
  운영 파일은 아무도 지키지 않고 있었다

고친 지점을 되돌려 세 테스트가 실제로 실패하는지 확인했다.

수명을 고쳐도 이미 발급된 토큰은 무효화되지 않는다. 신규 발급분만
10시간이 된다. 즉시 무효화가 필요하면 secret 교체가 따로 필요하고
전 사용자 재로그인을 동반한다.

+++ src/test/kotlin/com/pluxity/aiot/user/service/entity/DummyEntities.kt
@@ -18,7 +18,7 @@ fun dummyUser(
 fun dummyRefreshToken(
     username: String = "username",
     token: String = "token",
-    timeToLive: Int = 30,
+    timeToLive: Long = 30,
 ): RefreshToken = RefreshToken(username, token, timeToLive)
 
 fun dummyRole(
